+++ km_api/know_me/tests/views/test_km_user_detail_view.py
@@ -3,51 +3,53 @@ from rest_framework import status
 from know_me import serializers, views
 
 
-km_user_detail_view = views.KMUserDetailView.as_view()
+profile_detail_view = views.ProfileDetailView.as_view()
 
 
-def test_get_own_km_user(api_rf, km_user_factory, user_factory):
+def test_get_own_profile(api_rf, profile_factory):
     """
     A user should be able to get the details of their own Know Me user
     account.
     """
-    user = user_factory()
-    api_rf.user = user
+    profile = profile_factory()
+    km_user = profile.km_user
 
-    km_user = km_user_factory(user=user)
+    api_rf.user = km_user.user
 
-    request = api_rf.get(km_user.get_absolute_url())
-    response = km_user_detail_view(request, pk=km_user.pk)
-
-    serializer = serializers.KMUserDetailSerializer(
-        km_user,
-        context={'request': request})
+    request = api_rf.get(profile.get_absolute_url())
+    response = profile_detail_view(request, pk=profile.pk)
 
     assert response.status_code == status.HTTP_200_OK
+
+    serializer = serializers.ProfileDetailSerializer(
+        profile,
+        context={'request': request})
+
     assert response.data == serializer.data
 
 
-def test_update(api_rf, km_user_factory, user_factory):
+def test_update(api_rf, profile_factory):
     """
     Sending a PATCH request to the view should update the specified
     Know Me user.
     """
-    user = user_factory()
-    api_rf.user = user
+    profile = profile_factory(name='Old Name')
+    km_user = profile.km_user
 
-    km_user = km_user_factory(quote='Old Quote.', user=user)
+    api_rf.user = km_user.user
+
     data = {
-        'quote': 'New Quote.',
+        'name': 'New Name',
     }
 
-    request = api_rf.patch(km_user.get_absolute_url(), data)
-    response = km_user_detail_view(request, pk=km_user.pk)
+    request = api_rf.patch(profile.get_absolute_url(), data)
+    response = profile_detail_view(request, pk=profile.pk)
 
     assert response.status_code == status.HTTP_200_OK
 
-    km_user.refresh_from_db()
-    serializer = serializers.KMUserDetailSerializer(
-        km_user,
+    profile.refresh_from_db()
+    serializer = serializers.ProfileDetailSerializer(
+        profile,
         context={'request': request})
 
     assert response.data == serializer.data
