+++ km_api/know_me/tests/views/test_km_user_list_view.py
@@ -1,74 +1,53 @@
 from rest_framework import status
-from rest_framework.reverse import reverse
 
 from know_me import serializers, views
 
 
-km_user_list_view = views.KMUserListView.as_view()
-url = reverse('know-me:km-user-list')
+profile_list_view = views.ProfileListView.as_view()
 
 
-def test_create(api_rf, user_factory):
+def test_create(api_rf, km_user_factory):
     """
     Sending a POST request with valid data to the view should create a
-    new km_user for the user making the request.
+    new profile.
     """
-    user = user_factory()
-    api_rf.user = user
+    km_user = km_user_factory()
+
+    api_rf.user = km_user.user
 
     data = {
-        'name': user.get_short_name(),
-        'quote': "Hi, I'm {name}".format(name=user.get_short_name()),
+        'name': 'New Profile',
+        'is_default': True,
     }
 
-    request = api_rf.post(url, data)
-    response = km_user_list_view(request)
+    request = api_rf.post(km_user.get_profile_list_url(), data)
+    response = profile_list_view(request, pk=km_user.pk)
 
     assert response.status_code == status.HTTP_201_CREATED
 
-    serializer = serializers.KMUserListSerializer(
-        user.km_user,
+    serializer = serializers.ProfileListSerializer(
+        km_user.profiles.get(),
         context={'request': request})
 
     assert response.data == serializer.data
 
 
-def test_create_second_km_user(api_rf, km_user_factory, user_factory):
+def test_get_own(api_rf, profile_factory):
     """
     A user who already has an associated ``KMUser`` instance should not
     be able to create a second one.
     """
-    user = user_factory()
+    profile = profile_factory()
+    km_user = profile.km_user
+    user = km_user.user
+
     api_rf.user = user
 
-    km_user_factory(user=user)
+    request = api_rf.get(km_user.get_profile_list_url())
+    response = profile_list_view(request, pk=km_user.pk)
 
-    data = {
-        'name': user.get_short_name(),
-        'quote': "Hi, I'm {name}".format(name=user.get_short_name()),
-    }
-
-    request = api_rf.post(url, data)
-    response = km_user_list_view(request)
-
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-
-def test_get_own_km_user(api_rf, km_user_factory, user_factory):
-    """
-    If the requesting user has a km_user, then a GET request to this
-    view should contain that km_user.
-    """
-    user = user_factory()
-    api_rf.user = user
-
-    km_user = km_user_factory(user=user)
-
-    request = api_rf.get(url)
-    response = km_user_list_view(request)
-
-    serializer = serializers.KMUserListSerializer(
-        [km_user],
+    serializer = serializers.ProfileListSerializer(
+        km_user.profiles,
         context={'request': request},
         many=True)
